compressphotofast-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compressphotofast-cli"
version = "1.0.0"
description = "Cross-platform CLI tool for fast image compression"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "CompressPhotoFast Team"}
]
keywords = ["image", "compression", "cli", "jpeg", "png"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow>=10.0.0",
    "piexif>=1.1.3",
    "click>=8.1.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
]

[project.scripts]
compressphotofast = "src.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/compressphotofast-cli"
Repository = "https://github.com/yourusername/compressphotofast-cli"
